+++ package.json
@@ -12,6 +12,7 @@
     "@types/leaflet": "^1.9.21",
     "@types/react-calendar": "^3.9.0",
     "@types/recharts": "^1.8.29",
+    "babel-plugin-react-compiler": "^1.0.0",
     "browser-image-compression": "^2.0.2",
     "leaflet": "^1.9.4",
     "next": "16.1.6",
